requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
cffi==1.17.1
channels==4.2.2
constantly==23.10.4
cryptography==45.0.4
daphne==4.1.2 #3.0.2
Django==4.2.23 #4.2.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pkg_resources==0.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.13.2
zope.interface==7.2
